+++ vilmo/writer.go
@@ -1,4 +1,4 @@
-package appenddb
+package vilmo
 
 import (
 	"os"
